requirements.txt
Python==2.7.3
beautifulsoup4==4.1.3
httpie==0.3.0
nose==1.2.1
django-nose==1.1
requests==0.14.1
yolk==0.4.3
pep8==1.3.3
pyflakes==0.5.0
autopep8==0.8.1
Django==1.4.2
South==0.7.5
pycrypto==2.6
pytz
